requirements.txt
influxdb-client
numpy
pandas
pygooglenews
python-dotenv
redis
requests
schedule
transformers
torch
